[catalog] fold www into ckan-app
