+++ Makefile
@@ -1,6 +1,5 @@
 MOLECULE_SUITES := \
   software/ckan/catalog/harvest \
-  software/ckan/catalog/www \
   software/ckan/catalog/ckan-app \
   software/ckan/inventory \
   software/common/php-fixes
