+++ ansible/inventories/staging/group_vars/catalog-admin/vars.yml
@@ -1,0 +1,2 @@
+---
+catalog_ckan_allow_login_routes: true
