+++ ansible/roles/software/ckan/catalog/ckan-app/molecule/default/tests/test_default.py
@@ -87,3 +87,22 @@ def test_apache(host):
 
     assert apache.is_running
     assert apache.is_enabled
+
+
+def test_wsgi(host):
+    f = host.file('/etc/ckan/apache.wsgi')
+
+    assert f.exists
+    assert f.user == 'root'
+    assert f.group == 'www-data'
+    assert f.mode == 0o644
+
+
+def test_apache_site(host):
+    f = host.file('/etc/apache2/sites-enabled/ckan.conf')
+
+    assert f.exists
+    assert f.user == 'root'
+    assert f.group == 'www-data'
+    assert f.mode == 0o644
+    assert f.contains('ErrorLog /var/log/ckan/ckan.error.log')
