--- ansible/roles/software/ckan/catalog/www/.yamllint
@@ -1,10 +1,0 @@
-extends: default
-
-rules:
-  braces:
-    max-spaces-inside: 1
-    level: error
-  brackets:
-    max-spaces-inside: 1
-    level: error
-  line-length: disable
