--- ansible/roles/software/ckan/catalog/www/handlers/main.yml
@@ -1,3 +1,0 @@
----
-- name: reload apache2
-  service: name=apache2 state=reloaded
