--- ansible/roles/software/ckan/catalog/www/meta/main.yml
@@ -1,12 +1,0 @@
----
-galaxy_info:
-  author: Data.gov
-  description: catalog-web tasks
-  license: CC0-1.0
-  min_ansible_version: 2.0
-  platforms:
-    - name: Ubuntu
-      versions:
-        - trusty
-        - bionic
-dependencies: []
