--- ansible/roles/software/ckan/catalog/www/molecule/default/molecule.yml
@@ -1,22 +1,0 @@
----
-dependency:
-  name: galaxy
-driver:
-  name: docker
-lint:
-  name: yamllint
-platforms:
-  - name: catalog-www-trusty
-    image: ubuntu:trusty
-  - name: catalog-www-bionic
-    image: ubuntu:bionic
-provisioner:
-  name: ansible
-  lint:
-    name: ansible-lint
-scenario:
-  name: default
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
