--- ansible/roles/software/ckan/catalog/www/molecule/default/playbook.yml
@@ -1,5 +1,0 @@
----
-- name: Converge
-  hosts: all
-  roles:
-    - role: www
