--- ansible/roles/software/ckan/catalog/www/tasks/main.yml
@@ -1,23 +1,0 @@
----
-- name: Create log directory
-  file: >-
-    path={{ catalog_log_dir }}
-    state=directory
-    owner=root
-    group=www-data
-    mode=0755
-
-- name: Copy all needed files
-  copy: src={{ item }} dest=/{{ item }} mode=0644 owner=root group=www-data
-  with_items:
-    - etc/ckan/apache.wsgi
-  notify: reload apache2
-
-- name: Configure ckan.conf
-  template: >-
-    src=etc/apache2/sites-enabled/ckan.{{ readwrite_type|default('default') }}.conf.j2
-    dest=/etc/apache2/sites-enabled/ckan.conf
-    mode=0644
-    owner=root
-    group=www-data
-  notify: reload apache2
