--- ansible/roles/software/ckan/catalog/www/templates/etc/apache2/sites-enabled/ckan.default.conf.j2
@@ -1,68 +1,0 @@
-<VirtualHost 0.0.0.0:443>
-# TODO https://github.com/GSA/datagov-deploy/issues/570
-SSLEngine on
-</VirtualHost>
-
-<VirtualHost 0.0.0.0:80>
-    ServerName ckan
-
-    # CSW endpoints provided by pycsw
-    # based on request URI
-
-    # full CSW, all metadata
-    ProxyPass /csw-all http://localhost:8001
-
-    # default CSW, collection level metadata
-    ProxyPass /csw http://localhost:8000
-
-
-    WSGIScriptAlias / /etc/ckan/apache.wsgi
-
-    # Pass authorization info on (needed for rest api).
-    WSGIPassAuthorization On
-
-    # Deploy as a daemon (avoids conflicts between CKAN instances).
-    WSGIDaemonProcess ckan display-name=ckan processes={{ catalog_ckan_wsgi_processes }} threads=15
-
-    WSGIProcessGroup ckan
-
-    <Directory /etc/ckan>
-      Options All
-      AllowOverride All
-      Require all granted
-    </Directory>
-
-    ErrorLog {{ catalog_ckan_error_log }}
-    CustomLog {{ catalog_ckan_access_log }} combined
-
-    # fix Cross-Frame Scripting (XFS) vulnerability
-    Header set X-Frame-Options "SAMEORIGIN"
-
-    Header set Pragma "no-cache"
-    Header set Expires "-1"
-    Header set X-XSS-Protection "1; mode=block"
-    Header set X-Content-Type-Options "nosniff"
-
-    Header set Access-Control-Allow-Origin "*"
-    Header set Access-Control-Allow-Methods "POST, PUT, GET, DELETE, OPTIONS"
-    Header set Access-Control-Allow-Headers "X-CKAN-API-KEY, Authorization, Content-Type"
-
-    Header set Referrer-Policy "origin"
-
-    # Header set Content-Security-Policy "default-src 'self'"
-
-    RewriteEngine on
-
-    RewriteRule ^/api/rest/dataset$ api/action/package_search [R]
-    RewriteRule ^/api/1/rest/dataset$ api/3/action/package_search [R]
-    RewriteRule ^/api/2/rest/dataset$ api/3/action/package_search [R]
-
-    RewriteRule ^/api/action/package_list$ api/action/package_search [R]
-    RewriteRule ^/api/3/action/package_list$ api/3/action/package_search [R]
-
-    RewriteRule ^/api/action/current_package_list_with_resources$ api/action/package_search [R]
-    RewriteRule ^/api/3/action/current_package_list_with_resources$ api/3/action/package_search [R]
-
-    RewriteRule ^/api/3/action/resource_search$ api/action/package_search [R]
-    RewriteRule ^/api/action/resource_search$ api/action/package_search [R]
-</VirtualHost>
